% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btcmodel.R
\name{autoplot.btcmodel}
\alias{autoplot.btcmodel}
\title{Plot components for btc model}
\usage{
\method{autoplot}{btcmodel}(object, discarding = 0.05, title = NULL)
}
\arguments{
\item{object}{Object of class "btcmodel"}

\item{discarding}{percent for first part of the simulated chain
as burn in}

\item{title}{The main title (on top)}
}
\value{
None. Function produces a plot
}
\description{
Produces a plot of the trend, cycle 
and grow components from an btcmodel.
}

