% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-spec.R
\name{autoplot.spec.phase}
\alias{autoplot.spec.phase}
\title{Autoplotting Spectral Densities}
\usage{
\method{autoplot}{spec.phase}(x, ci = 0.95, xlab = "frequency",
  ylab = "phase", main = NULL)
}
\arguments{
\item{x}{an object of class "spec"}

\item{ci}{coverage probability for confidence interval. 
Plotting of the confidence bar/limits is omitted unless ci
is strictly positive.}

\item{xlab}{the x label of the plot}

\item{ylab}{the y label of the plot}

\item{main}{overall title for the plot. 
If missing, a suitable title is constructed}
}
\description{
Plotting method for objects of class "spec". For multivariate 
time series it plots the marginal spectra of the series or 
pairs plots of the coherency and phase of the cross-spectra.
}

