% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-spec.R
\name{autoplot.spec}
\alias{autoplot.spec}
\title{Autoplotting Spectral Densities}
\usage{
\method{autoplot}{spec}(x, ci = 0.95, log = TRUE, xlab = "frequency",
  ylab = NULL, main = NULL, plot.type = c("spectrum", "coherency",
  "phase"))
}
\arguments{
\item{x}{an object of class "spec"}

\item{ci}{coverage probability for confidence interval. 
Plotting of the confidence bar/limits is omitted unless ci
is strictly positive.}

\item{log}{use conventional log scale or linear scale. The default is TRUE.}

\item{xlab}{the x label of the plot}

\item{ylab}{the y label of the plot}

\item{main}{overall title for the plot. 
If missing, a suitable title is constructed}

\item{plot.type}{For multivariate time series, the type of plot required.}
}
\description{
Plotting method for objects of class "spec". For multivariate 
time series it plots the marginal spectra of the series or 
pairs plots of the coherency and phase of the cross-spectra.
}
