% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucwang.R
\name{autoplot.ucwang}
\alias{autoplot.ucwang}
\title{Plot components from ucwang model}
\usage{
\method{autoplot}{ucwang}(object)
}
\arguments{
\item{object}{Object of class "ucwanf"}
}
\value{
None. Function produces a plot
}
\description{
Produces a plot of the trend, cycle 
and grow components from an uc wang model.
}

