% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btcmodel.R
\name{btcmodel}
\alias{btcmodel}
\title{Modelo estrutural tendencia e ciclo bayesiano}
\usage{
btcmodel(y, a.theta, b.theta, dV = 1e-07, m0 = rep(0, 4), C0 = diag(x =
  c(rep(1e+07, 2), rep(1, 2))), n.sample = 1, thin = 0,
  save.states = FALSE)
}
\arguments{
\item{y}{data vector or univariate time series}

\item{a.theta}{prior mean of system precisions (recycled, if needed)}

\item{b.theta}{prior variance of system precisions (recycled, if needed)}

\item{dV}{the variance, or the diagonal elements of the variance 
matrix in the multivariate case, of the observation noise. 
V is assumed to be diagonal and it defaults to zero}

\item{m0}{the expected value of the pre-sample state vector}

\item{C0}{the variance matrix of the pre-sample state vector}

\item{n.sample}{requested number of Gibbs iterations}

\item{thin}{discard thin iterations for every saved iteration}

\item{save.states}{should the simulated states be included in the output?}
}
\value{
a list of Markov chain Monte Carlo (MCMC) simulated values.
 \item{phi}{simulated values of the AR parameter}
 \item{vars}{simulated values of the unknown variance}
 \item{theta}{simulated values of the state vectors}
}
\description{
Estima o modelo estutrural de componentes nao observados para
a tendencia estocastica e ciclo (gap) estacionario de forma 
bayesiana com MCMC e Gibbs sample. Para detalhes ver secao 4.6.1 
de Petris e Petrone (2009).
}
\references{
PETRIS, G.; PETRONE, S.; CAMPAGNOLI, P. 
Dynamic Linear Models with R. New York: Springer Science, 2009.
}

