% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{num2tab}
\alias{num2tab}
\title{Formata numeros para tabelas}
\usage{
num2tab(num, dig = 3)
}
\arguments{
\item{num}{vetor de numeros a serem formatados}

\item{dig}{quantidade de casas decimais (digitos usados apos a virgula)}
}
\value{
vetor dos numeros formatos em character
}
\description{
Formata numeros para tabelas e transforma para notacao centifica caso
as seja menor que tres casas decimais
}
