% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pibsetores.R
\docType{data}
\name{pibsetores}
\alias{pibsetores}
\title{PIB setorial do Brasil}
\format{Uma serie temporal mutivariada (mts) com 12 variaveis:
"Agropec.",	"Ext.Mineral",	"Transfor.",	"Eletricidade", "Construcao",	"Comercio",	"Transporte",	"Serv.infor.",	"Interm.finac.", 	"Ativ.imobili.",	"Outros Servicos" e	"Administacao publica"}
\source{
Instituto Brasileiro de Geografia e Estatistica (IBGE). Site www.ibge.gov.br
}
\usage{
pibsetores
}
\description{
Dados contem as series das contas trimestrais do PIB setorial de 1996Q1 ate 
2016Q3. Valores encadeados a preco de 1995 (1.000.000 Reais)
}
\keyword{datasets}

