% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btcmodel.R
\name{tab.btcmodel}
\alias{tab.btcmodel}
\title{Tabela resumo  btcmodel}
\usage{
tab.btcmodel(x, discarding = 0.05)
}
\arguments{
\item{x}{An object of class \link[mecnost]{btcmodel}}

\item{discarding}{percent for first part of the simulated chain
as burn in}
}
\value{
Uma tabela com as colunas
\item{Estimativas}{means of a matrix of simulated values}
\item{Desvio padrao}{estimate of the Monte Carlo standard deviation, obtained
using Sokal’s method}
\item{quantil 5}{0.05 quantiles of simulated values}
\item{quantil 95}{0.95 quantiles of simulated values}
}
\description{
Fornece a tabela resumo do coeficientes estimados por \link[mecnost]{btcmodel}
}
