% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucwang.R
\name{tab.ucwang}
\alias{tab.ucwang}
\title{Tabela resumo do modelo ucwang estimado}
\usage{
tab.ucwang(x)
}
\arguments{
\item{x}{An object of class \link[mecnost]{ucwang}}
}
\value{
Uma tabela com os coeficientes estimados
}
\description{
Fornece a tabela resumo do coeficientes estimados por \link[mecnost]{ucwang}
}
