% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucwang.R
\name{ucwang}
\alias{ucwang}
\title{Estima a tendencia e o ciclo com modelo estrutural de componentes nao observados}
\usage{
ucwang(x, init = c(1, 0.5, 0.2, 0.1, 0.1), model = 0, ...)
}
\arguments{
\item{x}{um objeto ts com a serie temporal}

\item{init}{valores iniciais da matriz de variancia dos estados, tendencia, ar1, ar2, tx crescimento.}

\item{model}{Model 0 general model, Model 1 where restrictions gc=0 and lambda=1 are imposed,
In Model 2, the mean value of the growth rate is simply gc.}

\item{...}{Optional arguments passed to function MARSS by control}
}
\value{
objeto da classe ucwang uma lista contendo o objeto retornado 
pela funcao (\link[MARSS]{MARSS}) e os dados x
}
\description{
Estima o modelo de componentes nao observaveis de Wang(2010) para obter a tendencia
e ciclo por maxima expectancia (EM) usando o pacote "MARSS"
}
\references{
WANG, P. An examination of business cycle features in UK Sectoral Output. Applied Economics, v. 42, n. 25, p. 3241–3252, 2010.
}
